Module ltn::filters

Re-exports

pub use self::existing::transform_existing_filters;

Modules

auto

Experiments to make a neighbourhood be low-traffic by automatically placing filters to prevent all shortcuts.

existing 🔒

Structs

ChangeKey

This logically changes every time an edit occurs. MapName isn’t captured here.

DiagonalFilter

A diagonal filter exists in an intersection. It’s defined by two roads (the order is arbitrary). When all of the intersection’s roads are sorted in clockwise order, this pair of roads splits the ordering into two groups. Turns in each group are still possible, but not across groups.

Edits

Stored in App per-map state. Before making any changes, call before_edit.

RoadFilter

A filter placed somewhere along a road

Toggle3Zoomed

Depending on the canvas zoom level, draws one of 2 things.

Enums

FilterType

Just determines the icon, has no semantics yet