Enum ltn::filters::auto::Heuristic

pub enum Heuristic {
    Greedy,
    BruteForce,
    SplitCells,
    OnlyOneBorder,
}

Variants

Greedy

Find the road with the most shortcuts that can be closed without creating a disconnected cell, and filter it.

There’s a vague intuition or observation that the “bottleneck” will have the most shortcuts going through it, so tackle the worst problem first.

BruteForce

Try adding one filter to every possible road, counting the shortcuts after. Choose the next step by the least resulting shortcuts.

SplitCells

Find one filter that splits a cell, maximizing the number of streets in each new cell.

OnlyOneBorder

Per cell, close all borders except for one. This doesn’t affect design_ltn, but prevents all shortcuts.

Implementations

impl Heuristic

pub fn choices() -> Vec<Choice<Heuristic>>

pub fn apply(
self,
ctx: &mut EventCtx<'_>,
app: &mut App,
neighbourhood: &Neighbourhood,
timer: &mut Timer<'_>
) -> Result<()>