Struct ltn::filters::DiagonalFilter

pub struct DiagonalFilter {
    r1: RoadID,
    r2: RoadID,
    i: IntersectionID,
    group1: BTreeSet<RoadID>,
    group2: BTreeSet<RoadID>,
}

A diagonal filter exists in an intersection. It’s defined by two roads (the order is arbitrary). When all of the intersection’s roads are sorted in clockwise order, this pair of roads splits the ordering into two groups. Turns in each group are still possible, but not across groups.

TODO Be careful with PartialEq! At a 4-way intersection, the same filter can be expressed as a different pair of two roads. And the (r1, r2) ordering is also arbitrary.

Fields
r1: RoadIDr2: RoadIDi: IntersectionIDgroup1: BTreeSet<RoadID>group2: BTreeSet<RoadID>
Implementations

impl DiagonalFilter

pub fn filters_for(
app: &SimpleApp<Session>,
i: IntersectionID
) -> Vec<DiagonalFilter>

Find all possible diagonal filters at an intersection

fn new(map: &Map, i: IntersectionID, r1: RoadID, r2: RoadID) -> DiagonalFilter

pub fn geometry(&self, map: &Map) -> Line

Physically where is the filter placed?

pub fn allows_turn(&self, from: RoadID, to: RoadID) -> bool

fn avoid_movements_between_roads(&self) -> Vec<(RoadID, RoadID)>