ltn

Module logic

Re-exports

pub use impact::Impact;

pub use turn_restrictions::possible_destination_roads;

Modules

auto_filters 🔒

Experiments to make a neighbourhood be low-traffic by automatically placing filters to prevent all shortcuts.

existing 🔒

impact

partition 🔒

shortcuts 🔒

turn_restrictions

Structs

BlockID

Identifies a single / unmerged block, which never changes

CustomBoundary

NeighbourhoodID

An opaque ID, won’t be contiguous as we adjust boundaries

Partitioning

Shortcuts

Enums

AutoFilterHeuristic

Functions

transform_existing

Edit the map, adding modal filters and crossings that’re modelled in OSM in various ways.