Struct ltn::neighbourhood::Neighbourhood

pub struct Neighbourhood {
    pub id: NeighbourhoodID,
    pub borders: BTreeSet<IntersectionID>,
    pub interior_intersections: BTreeSet<IntersectionID>,
    pub boundary_polygon: Polygon,
    pub interior_roads: BTreeSet<RoadID>,
    pub perimeter_roads: BTreeSet<RoadID>,
    pub suspicious_perimeter_roads: BTreeSet<RoadID>,
    pub connected_exterior_roads: BTreeSet<RoadID>,
    pub cells: Vec<Cell>,
    pub shortcuts: Shortcuts,
}

Fields
id: NeighbourhoodIDborders: BTreeSet<IntersectionID>

Intersections which form the boundary of the neighbourhood. This set includes any intersection which is connected to a road which is part of the neighbourhood’s perimeter. The roads which form the perimeter of the neighbourhood are the union of perimeter_roads and suspicious_perimeter_roads.

interior_intersections: BTreeSet<IntersectionID>

Intersections which are entirely inside the neighbourhood, and only connect interior roads to other interior roads.

boundary_polygon: Polygoninterior_roads: BTreeSet<RoadID>

Roads which are either (a) entirely inside the neighbourhood and (b) roads which are part of suspicious_perimeter_roads.

perimeter_roads: BTreeSet<RoadID>

Roads which form part of the neighbourhood’s perimeter, and are classified as arterial roads based on their OSM tags. suspicious_perimeter_roads are NOT included in perimeter_roads.

suspicious_perimeter_roads: BTreeSet<RoadID>

Roads which form part of the neighbourhood’s perimeter, but are classified as local roads based on their OSM tags. suspicious_perimeter_roads are always a subset of interior_roads.

connected_exterior_roads: BTreeSet<RoadID>

Roads which are lie outside the boundary_polygon but could potentially be connected to an interior_road or perimeter_road by either a road.turn_restrictions, or road.complicated_turn_restrictions. finish_init() populates this field.

cells: Vec<Cell>shortcuts: Shortcuts
Implementations

impl Neighbourhood

pub fn new(app: &App, id: NeighbourhoodID) -> Neighbourhood

pub fn new_without_app( map: &Map, partitioning: &Partitioning, id: NeighbourhoodID ) -> Neighbourhood

fn new_custom( map: &Map, id: NeighbourhoodID, custom: CustomBoundary ) -> Neighbourhood

fn finish_init(&mut self, map: &Map)

pub fn edits_changed(&mut self, map: &Map)

Recalculates cells and shortcuts after a relevant edit

pub fn fade_irrelevant(&self, ctx: &EventCtx<'_>, app: &App) -> Drawable