ltn::pages::design_ltn
Module turn_restrictions

Functions

build_focused_turns_geom 🔒

Builds the geom representing the FocusTurns hull and intersection

build_individual_target_road_geom 🔒

build_turn_options_geom 🔒

Builds the geom representing each of the individual turn options (permitted and restricted) within the FocusTurns.

build_turn_restriction_hover_geom 🔒

Builds the hover geom for showing turn restrictions when no FocusTurns are selected

handle_edited_turn_restrictions

handle_world_outcome

make_world

widget