Function ltn::per_neighborhood::populate_world

pub fn populate_world<T: ObjectID, F: Fn(FilterableObj) -> T>(
    ctx: &mut EventCtx<'_>, 
    app: &SimpleApp<Session>, 
    neighborhood: &Neighborhood, 
    world: &mut World<T>, 
    wrap_id: F, 
    zorder: usize
)

Adds clickable objects for managing filters on roads and intersections. The caller is responsible for base drawing behavior, initialize_hover, etc.