Struct ltn::Partitioning

pub struct Partitioning {
    pub map: MapName,
    neighborhoods: BTreeMap<NeighborhoodID, (Block, Color)>,
    single_blocks: Vec<Block>,
    neighborhood_id_counter: usize,
    block_to_neighborhood: BTreeMap<BlockID, NeighborhoodID>,
}

Fields
map: MapNameneighborhoods: BTreeMap<NeighborhoodID, (Block, Color)>single_blocks: Vec<Block>neighborhood_id_counter: usizeblock_to_neighborhood: BTreeMap<BlockID, NeighborhoodID>
Implementations

impl Partitioning

pub fn empty() -> Partitioning

Only valid before the LTN tool has been activated this session

pub fn seed_using_heuristics(
app: &SimpleApp<Session>,
timer: &mut Timer<'_>
) -> Partitioning

pub fn recalculate_coloring(&mut self) -> bool

True if the coloring changed

pub fn transfer_block(
&mut self,
map: &Map,
id: BlockID,
old_owner: NeighborhoodID,
new_owner: NeighborhoodID
) -> Result<Option<NeighborhoodID>>

pub fn remove_block_from_neighborhood(
&mut self,
map: &Map,
id: BlockID,
old_owner: NeighborhoodID
) -> Result<Option<NeighborhoodID>>

Needs to find an existing neighborhood to take the block, or make a new one

impl Partitioning

pub fn neighborhood_block(&self, id: NeighborhoodID) -> &Block

pub fn neighborhood_color(&self, id: NeighborhoodID) -> Color

pub fn all_neighborhoods(&self) -> &BTreeMap<NeighborhoodID, (Block, Color)>

fn neighborhood_containing(&self, find_block: BlockID) -> Option<NeighborhoodID>

pub fn all_single_blocks(&self) -> Vec<(BlockID, &Block)>

pub fn all_block_ids(&self) -> Vec<BlockID>

pub fn get_block(&self, id: BlockID) -> &Block

pub fn block_to_neighborhood(&self, id: BlockID) -> NeighborhoodID

pub fn calculate_frontier(&self, perim: &Perimeter) -> BTreeSet<BlockID>

Blocks on the “frontier” are adjacent to the perimeter, either just inside or outside.

fn make_merged_blocks(
&self,
map: &Map,
input: Vec<BlockID>
) -> Result<Vec<Block>>