Crate ltn

Modules

browse

colors

components

connectivity

draw_cells

export

filters

impact

neighborhood

partition

per_neighborhood

route_planner

save

select_boundary

shortcut_viewer

shortcuts

Structs

Args

BrowseNeighborhoods

Cell

A partitioning of the interior of a neighborhood based on driving connectivity

DiagonalFilter

A diagonal filter exists in an intersection. It’s defined by two roads (the order is arbitrary). When all of the intersection’s roads are sorted in clockwise order, this pair of roads splits the ordering into two groups. Turns in each group are still possible, but not across groups.

DistanceInterval

An interval along a road’s length, with start < end.

ModalFilters

Stored in App session state. Before making any changes, call before_edit.

Neighborhood

NeighborhoodID

An opaque ID, won’t be contiguous as we adjust boundaries

Partitioning

Session

Functions

after_edit

clear_current_proposal

draw_with_layering

Do the equivalent of SimpleApp::draw_unzoomed or draw_zoomed, but after the water/park areas layer, draw something custom.

main

run

Type Definitions

App

Transition